% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionTable.R
\docType{methods}
\name{anwartschaften}
\alias{anwartschaften}
\alias{anwartschaften,pensionTable-method}
\title{Calculates all "anwartschaften" for the gien pension table}
\usage{
anwartschaften(object, ...)

\S4method{anwartschaften}{pensionTable}(object, ..., i = 0.03, YOB = 1982)
}
\arguments{
\item{object}{A pension table object (instance of a \code{\linkS4class{pensionTable}} class)}

\item{...}{Currently unused}

\item{i}{Interest rate (default 0.03)}

\item{YOB}{Year of birth (default 1982)}
}
\description{
Calculates all "anwartschaften" for the gien pension table
}
\section{Methods (by class)}{
\itemize{
\item \code{pensionTable}: Calculates all "anwartschaften" for the gien pension table
}}

\examples{
pensionTables.load("Austria_*", wildcard=TRUE)
# anwartschaften(EttlPagler.male, i=0.03, YOB=1972)

}
