% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseTable.R
\docType{methods}
\name{baseTable}
\alias{baseTable}
\alias{baseTable,valuationTable-method}
\alias{baseTable,valuationTable.period-method}
\title{Return the base table of the life table}
\usage{
baseTable(object, ...)

\S4method{baseTable}{valuationTable}(object, ...)

\S4method{baseTable}{valuationTable.period}(object, ...)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the base table of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable}: Return the base table of the life table

\item \code{valuationTable.period}: Return the base table of the life table
}}

