% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseYear.R
\docType{methods}
\name{baseYear}
\alias{baseYear}
\alias{baseYear,valuationTable-method}
\alias{baseYear,valuationTable.mixed-method}
\title{Return the base year of the life table}
\usage{
baseYear(object, ...)

\S4method{baseYear}{valuationTable}(object, ...)

\S4method{baseYear}{valuationTable.mixed}(object, ...)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the base year of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable}: Return the base year of the life table

\item \code{valuationTable.mixed}: Return the base year of the life table
}}

