% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathProbabilities.R
\docType{methods}
\name{deathProbabilities}
\alias{deathProbabilities}
\alias{deathProbabilities,valuationTable.ageShift-method}
\alias{deathProbabilities,valuationTable.improvementFactors-method}
\alias{deathProbabilities,valuationTable.mixed-method}
\alias{deathProbabilities,valuationTable.period-method}
\alias{deathProbabilities,valuationTable.trendProjection-method}
\title{Return the (cohort) death probabilities of the life table given the birth year (if needed)}
\usage{
deathProbabilities(object, ..., YOB = 1975)

\S4method{deathProbabilities}{valuationTable.period}(object, ..., YOB = 1975)

\S4method{deathProbabilities}{valuationTable.ageShift}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{valuationTable.trendProjection}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{valuationTable.improvementFactors}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{valuationTable.mixed}(object, ..., YOB = 1975)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{...}{Other parameters (currently unused)}

\item{YOB}{The birth year for which the death probabilities should be calculated}
}
\description{
Return the (cohort) death probabilities of the life table given the birth year (if needed)
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable.period}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{valuationTable.ageShift}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{valuationTable.trendProjection}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{valuationTable.improvementFactors}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{valuationTable.mixed}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)
}}

