% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{fitExpExtrapolation}
\alias{fitExpExtrapolation}
\title{Fit an exponential function exp(-A*(x-x0)) to the last value (f(100) and f'(100) need to coincide):}
\usage{
fitExpExtrapolation(data, idx, up = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{data.frame to which an exponential function should be fit}

\item{idx}{Index of the position of the fit}

\item{up}{Whether the fit is forward- or backward-facing}

\item{verbose}{Whether to include data about the fit in the output}
}
\description{
Fit an exponential function exp(-A*(x-x0)) to the last value (f(100) and f'(100) need to coincide):
}
