% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmega.R
\docType{methods}
\name{getOmega}
\alias{getOmega}
\alias{getOmega,valuationTable.joined-method}
\alias{getOmega,valuationTable.mixed-method}
\alias{getOmega,valuationTable.observed-method}
\alias{getOmega,valuationTable.period-method}
\title{Return the maximum age of the life table}
\usage{
getOmega(object)

\S4method{getOmega}{valuationTable.period}(object)

\S4method{getOmega}{valuationTable.mixed}(object)

\S4method{getOmega}{valuationTable.joined}(object)

\S4method{getOmega}{valuationTable.observed}(object)
}
\arguments{
\item{object}{A life table object (instance of a \code{valuationTable} class)}
}
\description{
Return the maximum age of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable.period}: Return the maximum age of the period life table

\item \code{valuationTable.mixed}: Return the maximum age of the mixed life table

\item \code{valuationTable.joined}: Return the maximum age of the joined life table

\item \code{valuationTable.observed}: Return the maximum age of the joined life table
}}

