% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriodTable.R
\docType{methods}
\name{getPeriodTable}
\alias{getPeriodTable}
\alias{getPeriodTable,valuationTable-method}
\title{Return the period life table as a \code{valuationTable.period} object}
\usage{
getPeriodTable(object, Period, ...)

\S4method{getPeriodTable}{valuationTable}(object, Period, ...)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{Period}{The observation year, for which the death probabilities should
be determined}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the period life table as a \code{valuationTable.period} object
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable}: Return the period life table as a
\code{valuationTable.period} object
}}

