% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTable.R
\docType{methods}
\name{lifeTable}
\alias{lifeTable}
\alias{lifeTable,mortalityTable-method}
\title{Return the lifetable object (package lifecontingencies) for the cohort life table}
\usage{
lifeTable(object, ...)

\S4method{lifeTable}{mortalityTable}(object, ...)
}
\arguments{
\item{object}{The life table object (class inherited from mortalityTable)}

\item{...}{Parameters to be passed to the \code{deathProbabilities} method
of the life table}
}
\description{
Return the lifetable object (package lifecontingencies) for the cohort life table
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable}: Return the lifetable object (package lifecontingencies)
for the cohort life table
}}

