% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.R
\docType{class}
\name{mortalityTable-class}
\alias{mortalityTable-class}
\alias{mortalityTable}
\title{Class mortalityTable}
\description{
Class \code{mortalityTable} is the (virtual) base class for all mortality
tables. It contains the name and some general values applying to all
types of tables, but does not contain any data itself. Use a child class
to create actual mortality tables.
}
\section{Slots}{

\describe{
\item{\code{name}}{The human-readable name of the mortality table}

\item{\code{baseYear}}{The base year of the mortality table (e.g. for tables with trend projection)}

\item{\code{modification}}{A function that will be called with the final death probabilities
to give the user a way to modify the final probabilities}

\item{\code{loading}}{Additional security loading on the resulting table (single numeric
value, e.g. 0.05 adds 5\% security margin to the probabilities)}

\item{\code{data}}{Placeholder list to make it possible to store any kind of data
associated with the object inside the object (e.g. the underlying
data used to derive the death probabilities, parameters for adjustment, etc.)}
}}

