% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.deMoivre.R
\docType{class}
\name{mortalityTable.deMoivre-class}
\alias{mortalityTable.deMoivre-class}
\alias{mortalityTable.deMoivre}
\title{Class mortalityTable.deMoivre - Mortality table with de Moivre's law}
\description{
A period life table with maximum age omega dn the time of death identically
distributed on the interval [0, omega]. The only required slot is the maximum
age \code{omega}, all probabilities are calculated from it.
Optionally, a name and loading can be passed (inherited from
\code{\link{mortalityTable}}).
}
\section{Slots}{

\describe{
\item{\code{omega}}{Maximum age}
}}

\examples{
mm = mortalityTable.deMoivre(100)
plot(mm,
     mortalityTable.deMoivre(90),
     mortalityTable.deMoivre(50))

}
