% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.improvementFactors.R
\docType{class}
\name{mortalityTable.improvementFactors-class}
\alias{mortalityTable.improvementFactors-class}
\alias{mortalityTable.improvementFactors}
\title{Class mortalityTable.improvementFactors - Cohort life table with improvement
factors}
\description{
A cohort life table, obtained by an improvment factor projection
from a given base table (PODs for a given observation year).
}
\section{Slots}{

\describe{
\item{\code{baseYear}}{The base year for the improvements (\code{baseTable}
describes the death probabilities in this year)}

\item{\code{improvement}}{Yearly improvement factors per age}
}}

