% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.joined.R
\docType{class}
\name{mortalityTable.joined-class}
\alias{mortalityTable.joined}
\alias{mortalityTable.joined-class}
\title{Class mortalityTable.joined - Life table created by joining two life tables}
\description{
A cohort life table obtained by joining two cohort life tables, each of which
applies only to certain observation years (e.g. for the past use the observed
PoDs, and project them to the future with the trend projection)
}
\section{Slots}{

\describe{
\item{\code{table1}}{The first \code{mortalityTable}, valid for years given in \code{yearRange1}}

\item{\code{yearRange1}}{The years, for which \code{table1} describes the death probabilities}

\item{\code{table2}}{The second \code{mortalityTable}, valid for years given in \code{yearRange2}}

\item{\code{yearRange2}}{The years, for which \code{table2} describes the death probabilities}
}}

