% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.jointLives.R
\docType{class}
\name{mortalityTable.jointLives-class}
\alias{mortalityTable.jointLives}
\alias{mortalityTable.jointLives-class}
\title{Class mortalityTable.jointLives - Life table for multiple joint lives}
\description{
A cohort life table obtained by calculating joint death probabilities for
multiple lives, each possibly using a different mortality table.
}
\section{Slots}{

\describe{
\item{\code{table}}{The \code{mortalityTable} object for all lives (vector if different tables should be used for the different persons)}
}}

