% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.period.R
\docType{data}
\name{mortalityTable.NA}
\alias{mortalityTable.NA}
\title{Empty mortality table indicating NA}
\format{
An object of class \code{mortalityTable.period} of length 1.
}
\usage{
mortalityTable.NA
}
\description{
Empty mortality table indicating NA
}
\keyword{datasets}
