% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.observed.R
\docType{class}
\name{mortalityTable.observed-class}
\alias{mortalityTable.observed-class}
\alias{mortalityTable.observed}
\title{Class mortalityTable.observed - Life table from actual observations}
\description{
A cohort life table described by actual observations (data frame of PODs
per year and age)
}
\section{Slots}{

\describe{
\item{\code{deathProbs}}{The observed death probabilities (age-specific probability of dying within one year)}

\item{\code{years}}{The observation years}

\item{\code{ages}}{The observation ages}
}}

