% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.period.R
\name{mortalityTable.onceAndFuture}
\alias{mortalityTable.onceAndFuture}
\title{Generate a (deterministic) mortality table with all probabilities starting at a given age set to 1}
\usage{
mortalityTable.onceAndFuture(transitionAge,
  name = "Deterministic mortality table", ages = 0:99)
}
\arguments{
\item{transitionAge}{The age where the deterministic transition occurs}

\item{name}{The name of the table}

\item{ages}{The ages of the table}
}
\description{
Generate a (deterministic) mortality table with all probabilities starting at a given age set to 1
}
