% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.period.R
\docType{class}
\name{mortalityTable.period-class}
\alias{mortalityTable.period-class}
\alias{mortalityTable.period}
\title{Class mortalityTable.period - Period life tables}
\description{
A period life table, giving death probabilities for each age, up to
maximum age \code{omega}. The \code{baseYear} slot can be used to hold
information about the period.
}
\section{Slots}{

\describe{
\item{\code{ages}}{The ages corresponding to the entries of the deathProbs}

\item{\code{deathProbs}}{The one-year death probabilities for the ages}

\item{\code{exposures}}{(Optional) exposured used to determine death probabilities
(can be used as weights for smoothing, for variances, etc.)}
}}

