% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.period.R
\name{mortalityTable.zeroes}
\alias{mortalityTable.zeroes}
\title{Generate a mortality table with all probabilities set to zero.}
\usage{
mortalityTable.zeroes(name = "Zero mortality table", ages = 0:99)
}
\arguments{
\item{name}{The name of the table}

\item{ages}{The ages of the table}
}
\description{
Generate a mortality table with all probabilities set to zero.
}
