% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.R
\docType{package}
\name{MortalityTables-package}
\alias{MortalityTables}
\alias{MortalityTables-package}
\title{Provide life table classes for life insurance purposes}
\description{
Classes to implement, analyze and plot cohort life tables for actuarial calculations. Birth-year dependent cohort mortality tables using a yearly trend to extrapolate from a base year are implemented, as well as period life table, cohort life tables using an age shift, and merged life tables. Additionally, several data sets from various countries are included to provide widely-used tables out of the box.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.open-tools.net/R/r-mortality-tables}
  \item Report bugs at \url{https://gitlab.open-tools.net/R/r-mortality-tables/-/issues}
}

}
\author{
\strong{Maintainer}: Reinhold Kainhofer \email{reinhold@kainhofer.com} [copyright holder]

}
