% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.list.R
\name{mortalityTables.list}
\alias{mortalityTables.list}
\title{List all available sets of life tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing life table can then be loaded with \link{mortalityTables.load}.}
\usage{
mortalityTables.list(pattern = "*", package = c("MortalityTables",
  "MortalityTablesPrivate"), prefix = "MortalityTables")
}
\arguments{
\item{pattern}{Restrict the results only to life table sets that match the pattern (default: "*" to show all sets)}

\item{package}{The package that contains the desired dataset in its \code{extdata/}
directory. Defaults to the "MortalityTables" package.
Multiple packages can be given as a vector.}

\item{prefix}{The file prefix, defaults to MortalityTables. Can be overridden to list other types of files, like "PensionTables"}
}
\description{
List all available sets of life tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing life table can then be loaded with \link{mortalityTables.load}.
}
\examples{
mortalityTables.list()
mortalityTables.list("Austria_*")
mortalityTables.list("*Annuities")
mortalityTables.list(package = c("MyCustomPackage"))

}
