% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{mT.setName}
\alias{mT.setName}
\title{Sets a new name for the given mortality table or the list/table/array of mortalityTables}
\usage{
mT.setName(table, name)
}
\arguments{
\item{table}{A life table object (instance of a \code{mortalityTable} class) or a list, table or array of mortalityTable objects}

\item{name}{New name for the table.}
}
\description{
Sets a new name for the given mortality table or the list/table/array of mortalityTables
}
\examples{
mortalityTables.load("Austria_Annuities")
mT.setName(AVOe2005R.male, name = "Austrian male Annuity table 2005-R")

}
