% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.list.R
\name{pensionTables.list}
\alias{pensionTables.list}
\title{List all available sets of pension tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing pension table can then be loaded with \link{pensionTables.load}.}
\usage{
pensionTables.list(
  pattern = "*",
  package = c("^MortalityTables", "^PensionTables")
)
}
\arguments{
\item{pattern}{Restrict the results only to pension table sets that match the pattern with wildcards (default: "*" to show all sets)}

\item{package}{The package that contains the desired dataset in its \code{extdata/}
directory. Defaults to the "MortalityTables" package.
Multiple packages can be given as a vector, even using regular expressions.}
}
\description{
List all available sets of pension tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing pension table can then be loaded with \link{pensionTables.load}.
}
\examples{
pensionTables.list()
pensionTables.list("USA_*")
pensionTables.list(package = c("MyCustomPackage"))

}
