% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.load.R
\name{pensionTables.load}
\alias{pensionTables.load}
\title{Load a named set of pension tables provided by the \link{MortalityTables} package}
\usage{
pensionTables.load(dataset, package = c("^MortalityTables", "^PensionTables"))
}
\arguments{
\item{dataset}{The set of lifpensione tables to be loaded. A list of all available
data sets is provided by the function \code{\link{pensionTables.list}}.
Wildcards (*) are allowed to match and load multiple datasets.}

\item{package}{The package that contains the dataset in its \code{extdata/}
               directory. Defaults to all packages starting with names that
               start with "MortalityTables" or "PensionTables".
               Multiple packages can be given as a vector, even using regular expressions.

pensionTables.list()
pensionTables.load("*")
pensionTables.load("USA_PensionPlan_RP2014")}
}
\description{
Load a named set of pension tables provided by the \link{MortalityTables} package
}
