% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.jointLives.R
\name{periodDeathProbabilitiesIndividual}
\alias{periodDeathProbabilitiesIndividual}
\title{Return a matrix of the persons' individual period death probabilities of a
joint-life table (instance of \code{\link{mortalityTable.jointLives}})}
\usage{
periodDeathProbabilitiesIndividual(tables, period, ageDifferences)
}
\arguments{
\item{tables}{List of life table objects (object inherited from
\code{\link{mortalityTable}})}

\item{period}{The observation period}

\item{ageDifferences}{The age differences to the first person}
}
\description{
Return a matrix of the persons' individual period death probabilities of a
joint-life table (instance of \code{\link{mortalityTable.jointLives}})
}
\examples{
mortalityTables.load("Germany_Census")
periodDeathProbabilitiesIndividual(list(mort.DE.census.1924.26.male), 1977, c(0, 0))
periodDeathProbabilitiesIndividual(list(mort.DE.census.1924.26.male), 1977, c(0, -5, 13))

}
