% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionTable.R
\docType{methods}
\name{periodTransitionProbabilities}
\alias{periodTransitionProbabilities}
\alias{periodTransitionProbabilities,pensionTable-method}
\title{Return all period transition probabilities of the pension table}
\usage{
periodTransitionProbabilities(object, ...)

\S4method{periodTransitionProbabilities}{pensionTable}(object, Period = 2017,
  ..., as.data.frame = TRUE)
}
\arguments{
\item{object}{A pension table object (instance of a \code{\linkS4class{pensionTable}} class)}

\item{...}{Currently unused}

\item{Period}{Observation year}

\item{as.data.frame}{Whether the return value should be a data.frame or an array containing transition matrices}
}
\description{
Return all period transition probabilities of the pension table
}
\section{Methods (by class)}{
\itemize{
\item \code{pensionTable}: Return all transition probabilities of the pension table for the period Period
}}

\examples{
pensionTables.load("Austria_*", wildcard=TRUE)
# periodTransitionProbabilities(EttlPagler.male, Period = 2017)

}
