% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValuationTables.R
\name{plotValuationTables}
\alias{plotValuationTables}
\title{Plot multiple valuation tables (life tables) in one plot}
\usage{
plotValuationTables(data, ..., title = "", legend.position = c(0.9, 0.1),
  legend.key.width = unit(25, "mm"))
}
\arguments{
\item{data}{First life table to be plotted. Either a \code{data.frame} generated by \code{makeQxDataFrame} or a \code{valuationTable} object}

\item{...}{Additional life tables to be plotted (if \code{data} is a \code{valuationTable} object)}

\item{title}{The plot title}

\item{legend.position}{The position of the legend (default is \code{c(0.9,0.1)})}

\item{legend.key.width}{The keywith of the lines in the  legend (default is \code{unit(25,"mm")})}
}
\description{
\code{plotValuationTables} prints multiple life tables (objects of child classes of \code{valuationTable}) in one log-linear plot, with a legend showing the names of the tables.
}

