% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{PopulationData.AT2017}
\alias{PopulationData.AT2017}
\title{Austrian population count (exposure) and deaths in 2017}
\format{
A data frame holding female, male and total exposures as well as death
counts, indexed by age.
}
\source{
\url{http://www.mortality.org/}, \url{https://www.statistik.at/}
}
\usage{
data(PopulationData.AT2017)
}
\description{
Austrian population count (exposure) and deaths in 2017
}
\keyword{datasets}
