% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTable.R
\docType{class}
\name{valuationTable-class}
\alias{valuationTable}
\alias{valuationTable-class}
\title{Class valuationTable}
\description{
Class \code{valuationTable} is the (virtual) base class for all valuation
tables. It contains the name and some general values applying to all
types of tables, but does not contain any data itself. Use a child class
to create actual valuation tables.
}
\section{Slots}{

\describe{
\item{\code{name}}{The human-readable name of the valuation table}

\item{\code{baseYear}}{The base year of the valuation table (e.g. for tables with trend projection)}

\item{\code{modification}}{A function that will be called with the final death probabilities
to give the user a way to modify the final probabilities}

\item{\code{loading}}{Additional security loading on the resulting table (single numeric
value, e.g. 0.05 adds 5\% security margin to the probabilities)}
}}

