% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTable.mixed.R
\docType{class}
\name{valuationTable.mixed-class}
\alias{valuationTable.mixed}
\alias{valuationTable.mixed-class}
\title{Class valuationTable.mixed - Life table as a mix of two life tables}
\description{
A cohort life table obtained by mixing two life tables with the given weights
}
\section{Slots}{

\describe{
\item{\code{table1}}{The first \code{valuationTable}}

\item{\code{table2}}{The second \code{valuationTable}}

\item{\code{weight1}}{The weight of the first valuation table}

\item{\code{weight2}}{The weight of the second valuation table}

\item{\code{loading}}{Additional security loading}
}}

