% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTable.observed.R
\docType{class}
\name{valuationTable.observed-class}
\alias{valuationTable.observed}
\alias{valuationTable.observed-class}
\title{Class valuationTable.observed - Life table from actual observations}
\description{
A cohort life table described by actual observations (data frame of PODs
per year and age)
}
\section{Slots}{

\describe{
\item{\code{data}}{The observations}

\item{\code{years}}{The observation years}

\item{\code{ages}}{The observation ages}
}}

