% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTable.period.R
\docType{class}
\name{valuationTable.period-class}
\alias{valuationTable.period}
\alias{valuationTable.period-class}
\title{Class valuationTable.period - Period life tables}
\description{
A period life table, giving death probabilities for each age, up to
maximum age \code{omega}. The \code{baseYear} slot can be used to hold
information about the period.
}
\section{Slots}{

\describe{
\item{\code{ages}}{The ages corresponding to the entries of the deathProbs}

\item{\code{deathProbs}}{The one-year death probabilities for the ages}
}}

