% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_ageShift-class}
\alias{ageShift,valuationTable_ageShift-method}
\alias{deathProbabilities,valuationTable_ageShift-method}
\alias{periodDeathProbabilities,valuationTable_ageShift-method}
\alias{valuationTable_ageShift}
\alias{valuationTable_ageShift-class}
\title{A cohort life table, obtained by age-shifting from a given base table (death probabilities}
\usage{
ageShift(object, YOB = 1975, ...)

\S4method{ageShift}{valuationTable_ageShift}(object, YOB = 1975, ...)

\S4method{deathProbabilities}{valuationTable_ageShift}(object, ...,
  YOB = 1975)

\S4method{periodDeathProbabilities}{valuationTable_ageShift}(object, ...,
  Period = 1975)
}
\arguments{
\item{YOB}{The birth year for which the age shift should be determined.}

\item{Period}{The observation year for which the period death probabilities should be determined}

\item{YOB}{The birth year for which the age shift should be determined.}

\item{YOB}{The birth year for which the death probabilities should be calculated}
}
\description{
A cohort life table, obtained by age-shifting from a given base table (death probabilities

Return the age shift of the age-shifted life table given the birth year

Return the age shift of the age-shifted life table given the birth year

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year
}
\section{Slots}{

\describe{
\item{\code{ageShifts}}{A \code{data.frame} with columns \code{YOB} and \code{shifts} giving the age shifts for each birth year}
}}

