% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_improvementFactors-class}
\alias{deathProbabilities,valuationTable_improvementFactors-method}
\alias{periodDeathProbabilities,valuationTable_improvementFactors-method}
\alias{valuationTable_improvementFactors}
\alias{valuationTable_improvementFactors-class}
\title{A cohort life table, obtained by an improvment factor projection
from a given base table (PODs for a given observation year).}
\usage{
\S4method{deathProbabilities}{valuationTable_improvementFactors}(object, ...,
  YOB = 1975)

\S4method{periodDeathProbabilities}{valuationTable_improvementFactors}(object,
  ..., Period = 1975)
}
\arguments{
\item{YOB}{The birth year for which the death probabilities should be calculated}

\item{Period}{The observation year for which the period death probabilities should be determined}
}
\description{
A cohort life table, obtained by an improvment factor projection
from a given base table (PODs for a given observation year).

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year
}
\section{Slots}{

\describe{
\item{\code{baseYear}}{The base year for the improvements (\code{baseTable} describes the death probabilities in this year)}

\item{\code{improvement}}{Yearly improvement factors per age}
}}

