% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_joined-class}
\alias{ages,valuationTable_joined-method}
\alias{getOmega,valuationTable_joined-method}
\alias{valuationTable_joined}
\alias{valuationTable_joined-class}
\title{A cohort life table obtained by joining two cohort life tables, each of which
applies only to certain observation years (e.g. for the past use the observed
PODs, and project them to the future with the trend projection)}
\usage{
\S4method{getOmega}{valuationTable_joined}(object)

\S4method{ages}{valuationTable_joined}(object, ...)
}
\description{
A cohort life table obtained by joining two cohort life tables, each of which
applies only to certain observation years (e.g. for the past use the observed
PODs, and project them to the future with the trend projection)

Return the maximum age of the joined life table

Return the defined ages of the life table
}
\section{Slots}{

\describe{
\item{\code{table1}}{The first \code{valuationTable}, valid for years given in \code{yearRange1}}

\item{\code{yearRange1}}{The years, for which \code{table1} describes the death probabilities}

\item{\code{table2}}{The second \code{valuationTable}, valid for years given in \code{yearRange2}}

\item{\code{yearRange2}}{The years, for which \code{table2} describes the death probabilities}
}}

