% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_mixed-class}
\alias{ages,valuationTable_mixed-method}
\alias{baseYear,valuationTable_mixed-method}
\alias{deathProbabilities,valuationTable_mixed-method}
\alias{getOmega,valuationTable_mixed-method}
\alias{periodDeathProbabilities,valuationTable_mixed-method}
\alias{valuationTable_mixed}
\alias{valuationTable_mixed-class}
\title{A cohort life table obtained by mixing two life tables with the given weights}
\usage{
\S4method{getOmega}{valuationTable_mixed}(object)

\S4method{ages}{valuationTable_mixed}(object, ...)

\S4method{deathProbabilities}{valuationTable_mixed}(object, ..., YOB = 1975)

\S4method{periodDeathProbabilities}{valuationTable_mixed}(object, ...,
  Period = 1975)

\S4method{baseYear}{valuationTable_mixed}(object, ...)
}
\arguments{
\item{YOB}{The birth year for which the death probabilities should be calculated}

\item{Period}{The observation year for which the period death probabilities should be determined}
}
\description{
A cohort life table obtained by mixing two life tables with the given weights

Return the maximum age of the mixed life table

Return the defined ages of the life table

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year

Return the base year of the life table
}
\section{Slots}{

\describe{
\item{\code{table1}}{The first \code{valuationTable}}

\item{\code{table2}}{The second \code{valuationTable}}

\item{\code{weight1}}{The weight of the first valuation table}

\item{\code{weight2}}{The weight of the second valuation table}

\item{\code{loading}}{Additional security loading}
}}

