% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_observed-class}
\alias{valuationTable_observed}
\alias{valuationTable_observed-class}
\title{A cohort life table described by actual observations (data frame of PODs
per year and age)}
\description{
A cohort life table described by actual observations (data frame of PODs
per year and age)
}
\section{Slots}{

\describe{
\item{\code{data}}{The observations}
}}

