% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_period-class}
\alias{ages,valuationTable_period-method}
\alias{baseTable,valuationTable_period-method}
\alias{deathProbabilities,valuationTable_period-method}
\alias{getOmega,valuationTable_period-method}
\alias{periodDeathProbabilities,valuationTable_period-method}
\alias{valuationTable_period}
\alias{valuationTable_period-class}
\title{A period life table, giving death probabilities for each age, up to
maximum age omega. Optionally apply selection factors to the probabilities}
\usage{
\S4method{getOmega}{valuationTable_period}(object)

\S4method{ages}{valuationTable_period}(object, ...)

\S4method{deathProbabilities}{valuationTable_period}(object, ..., YOB = 1975)

\S4method{periodDeathProbabilities}{valuationTable_period}(object, ...)

\S4method{baseTable}{valuationTable_period}(object, ...)
}
\arguments{
\item{YOB}{The birth year for which the death probabilities should be calculated}

\item{Period}{The observation year for which the period death probabilities should be determined}
}
\description{
A period life table, giving death probabilities for each age, up to
maximum age omega. Optionally apply selection factors to the probabilities

Return the maximum age of the period life table

Return the defined ages of the life table

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year

Return the base table of the life table
}
\section{Slots}{

\describe{
\item{\code{ages}}{The ages corresponding to the entries of the deathProbs}

\item{\code{deathProbs}}{The one-year death probabilities for the ages}
}}

