% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable_trendProjection-class}
\alias{deathProbabilities,valuationTable_trendProjection-method}
\alias{periodDeathProbabilities,valuationTable_trendProjection-method}
\alias{undampenTrend,valuationTable_trendProjection-method}
\alias{valuationTable_trendProjection}
\alias{valuationTable_trendProjection-class}
\title{A cohort life table, obtained by a trend projection from a given base table
(PODs for a given observation year). Typically, the trend is obtained by
the Lee-Carter method or some other trend estimation.
The dampingFunction can be used to modify the cumulative years (e.g. G(tau+x) instead of tau+x)
If trend2 is given, the G(tau+x) gives the weight of the first trend, 1-G(tau+x) the weight of the second trend}
\usage{
\S4method{deathProbabilities}{valuationTable_trendProjection}(object, ...,
  YOB = 1975)

\S4method{periodDeathProbabilities}{valuationTable_trendProjection}(object, ...,
  Period = 1975)

undampenTrend(object)

\S4method{undampenTrend}{valuationTable_trendProjection}(object)
}
\arguments{
\item{YOB}{The birth year for which the death probabilities should be calculated}

\item{Period}{The observation year for which the period death probabilities should be determined}
}
\description{
A cohort life table, obtained by a trend projection from a given base table
(PODs for a given observation year). Typically, the trend is obtained by
the Lee-Carter method or some other trend estimation.
The dampingFunction can be used to modify the cumulative years (e.g. G(tau+x) instead of tau+x)
If trend2 is given, the G(tau+x) gives the weight of the first trend, 1-G(tau+x) the weight of the second trend

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year

Return a \code{valuationTable_trensProjection} object with the trend damping removed.

Return a \code{valuationTable_trensProjection} object with the trend damping removed.
}
\section{Slots}{

\describe{
\item{\code{baseYear}}{The base year of the trend projection (\code{baseTable} describes the death probabilities in this year)}

\item{\code{trend}}{The yearly improvements of the log-death probabilities (per age)}

\item{\code{dampingFunction}}{A possible damping of the trend. This is a function \code{damping(delta_years)} that gets a vector of years from the baseYear and should return the dampened values.}

\item{\code{trend2}}{The alternate trend. If given, the damping function interpolates between \code{trend} and \code{trend2}, otherwise the dumping function simply modifies the coefficients of \code{trend}.}
}}

