% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTable.R
\docType{package}
\name{ValuationTables-package}
\alias{ValuationTables}
\alias{ValuationTables-package}
\title{Provide life table classes for life insurance purposes}
\description{
Provide life table classes for life insurance purposes
}

