% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTables.list.R
\name{valuationTables.list}
\alias{valuationTables.list}
\title{List all available sets of life tables provided by the \link[ValuationTables]{ValuationTables-package} package
An existing life table can then be loaded with \link{valuationTables.load}.}
\usage{
valuationTables.list()
}
\description{
List all available sets of life tables provided by the \link[ValuationTables]{ValuationTables-package} package
An existing life table can then be loaded with \link{valuationTables.load}.
}

