% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittaker.mortalityTable.R
\name{whittaker.mortalityTable}
\alias{whittaker.mortalityTable}
\title{Smooth a life table using the Whittaker-Henderson method, intepolation possibly missing values}
\usage{
whittaker.mortalityTable(table, lambda = 10, d = 2,
  name.postfix = ", smoothed", ..., weights = NULL)
}
\arguments{
\item{table}{Mortality table to be graduated. Must be an instance of a
\code{mortalityTable}-derived class.}

\item{lambda}{Smoothing parameter (default 10)}

\item{d}{order of differences (default 2)}

\item{name.postfix}{Postfix appended to the name of the graduated table}

\item{...}{additional arguments (currently unused)}

\item{weights}{Vector of weights used for graduation. Entries with weight 0
will be interpolated. If not given, the exposures of the table
or equal weights are used.}

\item{reference}{The reference table that determines the 100\% values.
If not given, the absolute mortality values are
compared and plotted on a log-linear scale.}

\item{trend}{If set to \code{TRUE}, the function \code{\link{plotMortalityTrend}}
is used to plot the trends of the given tables.}
}
\description{
\code{whittaker.mortalityTable} uses the Whittaker-Henderson graduation method
to smooth a table of raw observed death probabilities, optionally using the
exposures stored in the table as weights (if no exposures are given, equal
weights are applied). All ages with a death probability of \code{NA} will be
interpolated in the Whittaker-Henderson method (see e.g. Lowrie)
}
\examples{
# TODO

}
\references{
Walter B. Lowrie: An Extension of the Whittaker-Henderson Method of Graduation, Transactions of Society of Actuaries, 1982, Vol. 34, pp. 329--372
}
\seealso{
\code{\link{whittaker}}
}
