% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages.R
\docType{methods}
\name{ages}
\alias{ages}
\alias{ages,mortalityTable.joined-method}
\alias{ages,mortalityTable.mixed-method}
\alias{ages,mortalityTable.observed-method}
\alias{ages,mortalityTable.period-method}
\title{Return the defined ages of the life table}
\usage{
ages(object, ...)

\S4method{ages}{mortalityTable.period}(object, ...)

\S4method{ages}{mortalityTable.mixed}(object, ...)

\S4method{ages}{mortalityTable.joined}(object, ...)

\S4method{ages}{mortalityTable.observed}(object, ...)
}
\arguments{
\item{object}{A life table object (instance of a \code{\linkS4class{mortalityTable}} class)}

\item{...}{Currently unused}
}
\description{
Return the defined ages of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable.period}: Return the defined ages of the period life table

\item \code{mortalityTable.mixed}: Return the defined ages of the mixed life table

\item \code{mortalityTable.joined}: Return the defined ages of the joined life table

\item \code{mortalityTable.observed}: Return the defined ages of the observed life table
}}
\examples{
mortalityTables.load("Austria_*", wildcard=TRUE)
ages(AVOe2005R.male)
ages(AVOe1996R.male)
ages(mort.AT.census.2011.male)

}

