% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageShift.R
\docType{methods}
\name{ageShift}
\alias{ageShift}
\alias{ageShift,valuationTable.ageShift-method}
\title{Return the age shift of the age-shifted life table given the birth year}
\usage{
ageShift(object, YOB = 1975, ...)

\S4method{ageShift}{valuationTable.ageShift}(object, YOB = 1975, ...)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{YOB}{The birth year for which the age shift should be determined.}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the age shift of the age-shifted life table given the birth year
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable.ageShift}: Return the age shift of the age-shifted life table
given the birth year
}}

