% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseTable.R, R/mortalityTable.jointLives.R
\docType{methods}
\name{baseTable}
\alias{baseTable}
\alias{baseTable,mortalityTable-method}
\alias{baseTable,mortalityTable.period-method}
\alias{baseTable,mortalityTable.jointLives-method}
\title{Return the base table of the life table}
\usage{
baseTable(object, ...)

\S4method{baseTable}{mortalityTable}(object, ...)

\S4method{baseTable}{mortalityTable.period}(object, ...)

\S4method{baseTable}{mortalityTable.jointLives}(object, ...)
}
\arguments{
\item{object}{The life table object (class inherited from mortalityTable)}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the base table of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable}: Return the base table of the life table

\item \code{mortalityTable.period}: Return the base table of the life table

\item \code{mortalityTable.jointLives}: Return the base table of the joint lives mortality table (returns the base table of the first table used for joint lives)
}}

\examples{
mortalityTables.load("Austria_Annuities")
baseTable(AVOe2005R.male)

}
