% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseYear.R, R/mortalityTable.jointLives.R
\name{baseYear}
\alias{baseYear}
\alias{baseYear,mortalityTable-method}
\alias{baseYear,mortalityTable.mixed-method}
\alias{baseYear,mortalityTable.jointLives-method}
\title{Return the base year of the life table}
\usage{
baseYear(object, ...)

\S4method{baseYear}{mortalityTable}(object, ...)

\S4method{baseYear}{mortalityTable.mixed}(object, ...)

\S4method{baseYear}{mortalityTable.jointLives}(object, ...)
}
\arguments{
\item{object}{The life table object (class inherited from mortalityTable)}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the base year of the life table
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable}: Return the base year of the life table

\item \code{mortalityTable.mixed}: Return the base year of the life table

\item \code{mortalityTable.jointLives}: Return the base year of the life table
}}

\examples{
mortalityTables.load("Austria_Annuities")
baseYear(AVOe2005R.male)

}
