% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathProbabilities.R
\docType{methods}
\name{deathProbabilities}
\alias{deathProbabilities}
\alias{deathProbabilities,mortalityTable.ageShift-method}
\alias{deathProbabilities,mortalityTable.improvementFactors-method}
\alias{deathProbabilities,mortalityTable.mixed-method}
\alias{deathProbabilities,mortalityTable.period-method}
\alias{deathProbabilities,mortalityTable.trendProjection-method}
\title{Return the (cohort) death probabilities of the life table given the birth year (if needed)}
\usage{
deathProbabilities(object, ..., YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.period}(object, ..., YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.ageShift}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.trendProjection}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.improvementFactors}(object, ...,
  YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.mixed}(object, ..., YOB = 1975)
}
\arguments{
\item{object}{The life table object (class inherited from mortalityTable)}

\item{...}{Other parameters (currently unused)}

\item{YOB}{The birth year for which the death probabilities should be calculated}
}
\description{
Return the (cohort) death probabilities of the life table given the birth year (if needed)
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable.period}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{mortalityTable.ageShift}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{mortalityTable.trendProjection}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{mortalityTable.improvementFactors}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{mortalityTable.mixed}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)
}}

