% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathProbabilities.R,
%   R/mortalityTable.jointLives.R, R/mortalityTable.observed.R
\name{deathProbabilities}
\alias{deathProbabilities}
\alias{deathProbabilities,mortalityTable.period-method}
\alias{deathProbabilities,mortalityTable.ageShift-method}
\alias{deathProbabilities,mortalityTable.trendProjection-method}
\alias{deathProbabilities,mortalityTable.improvementFactors-method}
\alias{deathProbabilities,mortalityTable.mixed-method}
\alias{deathProbabilities,mortalityTable.jointLives-method}
\alias{deathProbabilities,mortalityTable.observed-method}
\title{Return the (cohort) death probabilities of the life table given the birth year (if needed)}
\usage{
deathProbabilities(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.period}(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.ageShift}(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.trendProjection}(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.improvementFactors}(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.mixed}(object, ..., ages = NULL, YOB = 1975)

\S4method{deathProbabilities}{mortalityTable.jointLives}(
  object,
  ...,
  agesJoint = NULL,
  ageDifferences = NULL,
  ages = NULL,
  YOB = 1975
)

\S4method{deathProbabilities}{mortalityTable.observed}(object, ..., ages = NULL, YOB = 1975)
}
\arguments{
\item{object}{The life table object (class inherited from mortalityTable)}

\item{...}{Other parameters (currently unused)}

\item{ages}{The vector of desired ages in the return value. This refers to the
age of the first person in the joint-life status. The ages of the
other persons will follow the given ageDifferences or YOB or age
vectors}

\item{YOB}{The vector of birth-years, or the birth-year of the first person.}

\item{ageDifferences}{A vector of age differences (relative to the first person) of all joint lives.}

\item{age}{The vector of ages of the persons under consideration}
}
\description{
Return the (cohort) death probabilities of the life table given the birth year (if needed)
}
\details{
The ages and years of birth of the joint lives can be given in multiple arguments. They will be evaluated in the following order:
\itemize{
  \item \code{YOB}: vector of birth years for all joint lives
  \item \code{YOB} (scalar, for the first life) and \code{ageDifferences}: relative to the first life
  \item \code{YOB} (scalar, for the first life) and \code{agesJoint}: explicit list of all ages of the joint lives
  \item \code{ages} (vector of requested ages, for the first life ONLY) and \code{ageDifferences}: relative to the first life
  \item \code{agesJoint}: numeric vector of individual ages
  \item \code{ageDifferences} only (not permitted for cohort life tables)
}
}
\section{Methods (by class)}{
\itemize{
\item \code{deathProbabilities(mortalityTable.period)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.ageShift)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.trendProjection)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.improvementFactors)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.mixed)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.jointLives)}: Return the (cohort) death probabilities of the life table given the birth year (if needed)

\item \code{deathProbabilities(mortalityTable.observed)}: Return the (cohort) death probabilities of the
life table given the birth year (if needed)

}}
\examples{
mortalityTables.load("Austria_Annuities")
deathProbabilities(AVOe2005R.male, YOB = 1975)
deathProbabilities(AVOe2005R.male, YOB = 2017)

mortalityTables.load("Germany_Census")
table.JL = mortalityTable.jointLives(
    name = "ADSt 24/26 auf verbundene Leben",
    table = mort.DE.census.1924.26.male
)
deathProbabilities(table.JL, YOB = 1977, ageDifferences = c(1, 5, -5, 16))
deathProbabilities(table.JL, YOB = 1977, ageDifferences = c(0))
deathProbabilities(table.JL, YOB = 1977, ageDifferences = c(1, 5, 16))

}
