% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.jointLives.R
\name{deathProbabilitiesIndividual}
\alias{deathProbabilitiesIndividual}
\title{Return a matrix of the persons' individual death probabilities of a joint-life
table (instance of \code{\link{mortalityTable.jointLives}})}
\usage{
deathProbabilitiesIndividual(tables, YOB = NULL, ageDifferences = NULL)
}
\arguments{
\item{tables}{List of life table objects (object inherited from
\code{\link{mortalityTable}})}

\item{YOB}{The birth year for the first person}

\item{ageDifferences}{The age differences to the first person}
}
\description{
The rows will always be the full age range of the first life, i.e. the first
row describes the mortality of the joint-life state when the first person is
in its first year of life. The last row will be for the maximum age of the first
life, even if other lives have shorter tables. These will be padded with the
last available value (which for life tables should be q_x=1 anyway).
}
\examples{
mortalityTables.load("Germany_Census")
deathProbabilitiesIndividual(list(mort.DE.census.1924.26.male), 1977, c(0, 0))
deathProbabilitiesIndividual(list(mort.DE.census.1924.26.male), 1977, c(0, -5, 13))

}
