% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortTable.R
\docType{methods}
\name{getCohortTable}
\alias{getCohortTable}
\alias{getCohortTable,valuationTable-method}
\title{Return the cohort life table as a \code{valuationTable.period} object}
\usage{
getCohortTable(object, YOB, ...)

\S4method{getCohortTable}{valuationTable}(object, YOB, ...)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{YOB}{The birth year for which the life table should be calculated}

\item{...}{Other parameters (currently unused)}
}
\description{
Return the cohort life table as a \code{valuationTable.period} object
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable}: Return the cohort life table as a
\code{valuationTable.period} object
}}

