% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.ageShift.R
\docType{class}
\name{mortalityTable.ageShift-class}
\alias{mortalityTable.ageShift-class}
\alias{mortalityTable.ageShift}
\title{Class mortalityTable.ageShift - Cohort life tables generated by age-shift}
\description{
A cohort life table, obtained by age-shifting from a given base table (death probabilities
}
\section{Slots}{

\describe{
\item{\code{ageShifts}}{A \code{data.frame} with columns \code{YOB} and \code{shifts} giving the age shifts for each birth year}
}}

