% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.mixed.R
\docType{class}
\name{mortalityTable.mixed-class}
\alias{mortalityTable.mixed}
\alias{mortalityTable.mixed-class}
\title{Class mortalityTable.mixed - Life table as a mix of two life tables}
\description{
A cohort life table obtained by mixing two life tables with the given weights
}
\section{Slots}{

\describe{
\item{\code{table1}}{The first \code{mortalityTable}}

\item{\code{table2}}{The second \code{mortalityTable}}

\item{\code{weight1}}{The weight of the first mortality table}

\item{\code{weight2}}{The weight of the second mortality table}

\item{\code{loading}}{Additional security loading}
}}

