% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.R
\docType{package}
\name{MortalityTables-package}
\alias{MortalityTables}
\alias{MortalityTables-package}
\title{Provide life table classes for life insurance purposes}
\description{
Provide life table classes for life insurance purposes
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.open-tools.net/R/r-mortality-tables}
}

}
\author{
\strong{Maintainer}: Reinhold Kainhofer \email{reinhold@kainhofer.com}

}
