% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTable.R
\docType{package}
\name{MortalityTables-package}
\alias{MortalityTables}
\alias{MortalityTables-package}
\title{Provide life table classes for life insurance purposes}
\description{
Provide life table classes for life insurance purposes
}

