% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.list.R
\name{mortalityTables.list}
\alias{mortalityTables.list}
\title{List all available sets of life tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing life table can then be loaded with \link{mortalityTables.load}.}
\usage{
mortalityTables.list(pattern = "*", package = "MortalityTables")
}
\arguments{
\item{pattern}{Restrict the results only to life table sets that match the pattern (default: "*" to show all sets)}

\item{package}{The package that contains the desired dataset in its \code{extdata/}
directory. Defaults to the "MortalityTables" package.}
}
\description{
List all available sets of life tables provided by the \link[MortalityTables]{MortalityTables-package} package
An existing life table can then be loaded with \link{mortalityTables.load}.
}
