% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.load.R
\name{mortalityTables.load}
\alias{mortalityTables.load}
\title{Load a named set of mortality tables provided by the \link{MortalityTables} package}
\usage{
mortalityTables.load(dataset, wildcard = FALSE)
}
\arguments{
\item{dataset}{The set of life tables to be loaded. A list of all available
data sets is provided by the function \code{\link{mortalityTables.list}}.}

\item{wildcard}{Whether the dataset name contains wildcard. If TRUE, all
datasets matching the pattern will be loaded}
}
\description{
Load a named set of mortality tables provided by the \link{MortalityTables} package
}

