% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionTable.R
\docType{class}
\name{pensionTable-class}
\alias{pensionTable-class}
\alias{pensionTable}
\title{Class pensionTable}
\description{
Class \code{pensionTable} is the (virtual) base class for all pensions
tables. It contains the name and some general values applying to all
types of tables. In particular, it holds individual tables for each of the
transition probabilities. Possible states are:
    * active: healty, no pension, typically paying some kin of premium
    * incapacity: disablity pension, in most cases permanent, not working, early pension
    * retirement: old age pension, usually starting with a fixed age
    * dead
      * Widow/widower pension
}
\details{
Correspondingly, the following transition probabilities can be given:
    * qxaa:  death probability of actives (active -> dead)
    * ix:    invalidity probability (active -> incapacity)
    * qxi:   death probability of invaid (invalid -> dead)
    * rx:    reactivation probability (incapacity -> active)
    * apx:   retirement probability (active -> retirement), typically 1 for a fixed age
    * apx:   retirement probability of invalids (invalid -> retirement), typically 0 or 1 for a fixed age
    * qxApm: death probability of retired (retired -> dead)
    * hx:    probability of a widow at moment of death (dead -> widow), y(x) age differene
    * qxw:   death probability of widows/widowers
}
\section{Slots}{

\describe{
\item{\code{qx}}{Death probability table of actives (derived from mortalityTable)}

\item{\code{ix}}{Invalidity probability of actives (derived from mortalityTable)}

\item{\code{qxi}}{Death probability table of invalids (derived from mortalityTable)}

\item{\code{rx}}{Reactivation probability of invalids (derived from mortalityTable)}

\item{\code{apx}}{Retirement probability of actives (derived from mortalityTable)}

\item{\code{apxi}}{Retirement probability of invalids (derived from mortalityTable)}

\item{\code{qxApm}}{Death probability of old age pensioners (derived from mortalityTable)}

\item{\code{hx}}{Probability of a widow at the moment of death (derived from mortalityTable)}

\item{\code{qxw}}{Death probability of widow(er)s (derived from mortality Table)}

\item{\code{xy}}{Age difference of the widow to the deceased}
}}

