% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalityTables.load.R
\name{pensionTables.load}
\alias{pensionTables.load}
\title{Load a named set of pension tables provided by the \link{MortalityTables} package}
\usage{
pensionTables.load(dataset, wildcard = FALSE, package = "MortalityTables")
}
\arguments{
\item{dataset}{The set of lifpensione tables to be loaded. A list of all available
data sets is provided by the function \code{\link{pensionTables.list}}.}

\item{wildcard}{Whether the dataset name contains wildcard. If TRUE, all
datasets matching the pattern will be loaded}

\item{package}{The package that contains the dataset in its \code{extdata/}
directory. Defaults to the "MortalityTables" package.}
}
\description{
Load a named set of pension tables provided by the \link{MortalityTables} package
}
