% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodDeathProbabilities.R
\docType{methods}
\name{periodDeathProbabilities}
\alias{periodDeathProbabilities}
\alias{periodDeathProbabilities,valuationTable.ageShift-method}
\alias{periodDeathProbabilities,valuationTable.improvementFactors-method}
\alias{periodDeathProbabilities,valuationTable.mixed-method}
\alias{periodDeathProbabilities,valuationTable.period-method}
\alias{periodDeathProbabilities,valuationTable.trendProjection-method}
\title{Return the (period) death probabilities of the life table for a given
observation year}
\usage{
periodDeathProbabilities(object, ..., Period = 1975)

\S4method{periodDeathProbabilities}{valuationTable.period}(object, ...,
  Period = 1975)

\S4method{periodDeathProbabilities}{valuationTable.ageShift}(object, ...,
  Period = 1975)

\S4method{periodDeathProbabilities}{valuationTable.trendProjection}(object, ...,
  Period = 1975)

\S4method{periodDeathProbabilities}{valuationTable.improvementFactors}(object,
  ..., Period = 1975)

\S4method{periodDeathProbabilities}{valuationTable.mixed}(object, ...,
  Period = 1975)
}
\arguments{
\item{object}{The life table object (class inherited from valuationTable)}

\item{...}{Other parameters (currently unused)}

\item{Period}{The observation year for which the period death probabilities should be determined}
}
\description{
Return the (period) death probabilities of the life table for a given
observation year
}
\section{Methods (by class)}{
\itemize{
\item \code{valuationTable.period}: Return the (period) death probabilities
of the life table for a given observation year

\item \code{valuationTable.ageShift}: Return the (period) death probabilities
of the life table for a given observation year

\item \code{valuationTable.trendProjection}: Return the (period) death probabilities
of the life table for a given observation year

\item \code{valuationTable.improvementFactors}: Return the (period) death probabilities
of the life table for a given observation year

\item \code{valuationTable.mixed}: Return the (period) death probabilities
of the life table for a given observation year
}}

