% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{PopulationData.AT2017}
\alias{PopulationData.AT2017}
\title{Austrian population count (exposure) and deaths in 2017}
\format{
A data frame holding female, male and total exposures as well as death
counts, indexed by age.
}
\source{
\url{https://www.mortality.org/}, \url{https://www.statistik.at/}
}
\usage{
data(PopulationData.AT2017)
}
\description{
This data.frame hold the official population counts (in terms of exposure)
as well as the death counts for Austria during the calendar year 2017.
}
\details{
The data was downloaded from \url{https://www.mortality.org/}, where is was
submitted by Statistik Austria.
}
\keyword{datasets}
