% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setModification.R
\docType{methods}
\name{setModification}
\alias{setModification}
\alias{setModification,mortalityTable-method}
\title{Return a copy of the table with the given modification function added}
\usage{
setModification(object, modification = 0)

\S4method{setModification}{mortalityTable}(object, modification = 0)
}
\arguments{
\item{object}{A life table object (instance of a \code{mortalityTable} class)}

\item{modification}{The postprocessing modification function (for example, so enforce a lower bound).}
}
\description{
Return a copy of the table with the given modification function added
}
\section{Methods (by class)}{
\itemize{
\item \code{mortalityTable}: Return the life table with the given modification set
}}

