% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionTable.R
\name{transitionProbabilities}
\alias{transitionProbabilities}
\title{Return all transition probabilities of the pension table}
\usage{
transitionProbabilities(object, ...)
}
\arguments{
\item{object}{A pension table object (instance of a \code{\linkS4class{pensionTable}} class)}

\item{...}{Currently unused}

\item{YOB}{Year of birth}
}
\description{
Return all transition probabilities of the pension table
}
\examples{
pensionTables.load("Austria_*", wildcard=TRUE)
transitionProbabilities(EttlPagler.male)

}
