% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensionTable.R
\docType{methods}
\name{transitionProbabilities}
\alias{transitionProbabilities}
\alias{transitionProbabilities,pensionTable-method}
\title{Return all transition probabilities of the pension table (generational probabilities)}
\usage{
transitionProbabilities(object, ...)

\S4method{transitionProbabilities}{pensionTable}(object, YOB = 1982, ...,
  as.data.frame = TRUE)
}
\arguments{
\item{object}{A pension table object (instance of a \code{\linkS4class{pensionTable}} class)}

\item{...}{Currently unused}

\item{YOB}{Year of birth}
}
\description{
Return all transition probabilities of the pension table (generational probabilities)
}
\section{Methods (by class)}{
\itemize{
\item \code{pensionTable}: Return all transition probabilities of the pension table for the generation YOB
}}

\examples{
pensionTables.load("Austria_*", wildcard=TRUE)
# transitionProbabilities(EttlPagler.male)

}
