% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValuationTables.R
\docType{class}
\name{valuationTable-class}
\alias{baseTable,valuationTable-method}
\alias{baseYear,valuationTable-method}
\alias{getCohortTable,valuationTable-method}
\alias{getPeriodTable,valuationTable-method}
\alias{lifeTable,valuationTable-method}
\alias{valuationTable}
\alias{valuationTable-class}
\title{Class valuationTable}
\usage{
getOmega(object)

ages(object, ...)

deathProbabilities(object, ..., YOB = 1975)

periodDeathProbabilities(object, ...)

lifeTable(object, ...)

\S4method{lifeTable}{valuationTable}(object, ...)

baseYear(object, ...)

\S4method{baseYear}{valuationTable}(object, ...)

baseTable(object, ...)

\S4method{baseTable}{valuationTable}(object, ...)

getPeriodTable(object, Period, ...)

\S4method{getPeriodTable}{valuationTable}(object, Period, ...)

getCohortTable(object, YOB, ...)

\S4method{getCohortTable}{valuationTable}(object, YOB, ...)
}
\arguments{
\item{...}{Parameters to be handed to the \code{deathProbabilities} method of the life table}

\item{YOB}{The birth year for which the death probabilities should be calculated}

\item{Period}{The observation year for which the period death probabilities should be determined}

\item{...}{Parameters to be handed to the \code{deathProbabilities} method of the life table}

\item{Period}{The observation year, for which the death probabilities should be determined}

\item{Period}{The observation year, for which the death probabilities should be determined}

\item{YOB}{The birth year for which the life table should be calculated}

\item{YOB}{The birth year for which the life table should be calculated}
}
\description{
Class \code{valuationTable} is the (virtual) base class for all valuation
tables. It contains the name and some general values applying to all
types of tables, but does not contain any data itself. Use a child class
to create actual valuation tables.

Return the maximum age of the life table

Return the defined ages of the life table

Return the (cohort) death probabilities of the life table given the birth year (if needed)

Return the (period) death probabilities of the life table for a given observation year

Return the lifetable object (package lifecontingencies) for the cohort life table

Return the lifetable object (package lifecontingencies) for the cohort life table

Return the base year of the life table

Return the base year of the life table

Return the base table of the life table

Return the base table of the life table

Return the period life table as a \code{valuationTable_period} object

Return the period life table as a \code{valuationTable_period} object

Return the cohort life table as a \code{valuationTable_period} object

Return the cohort life table as a \code{valuationTable_period} object
}
\section{Slots}{

\describe{
\item{\code{name}}{The human-readable name of the valuation table}

\item{\code{baseYear}}{The base year of the valuation table (e.g. for tables with trend projection)}

\item{\code{modification}}{A function that will be called with the final death probabilities
to give the user a way to modify the final probabilities}

\item{\code{loading}}{Additional security loading on the resulting table (single numeric
value, e.g. 0.05 adds 5% security margin to the probabilities)}
}}

