% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuationTables.load.R
\name{valuationTables.load}
\alias{valuationTables.load}
\title{Load a named set of life tables provided by the \link{ValuationTables} package}
\usage{
valuationTables.load(dataset)
}
\arguments{
\item{dataset}{The set of life tables to be loaded. A list of all available data sets is provided by the}
}
\description{
Load a named set of life tables provided by the \link{ValuationTables} package
}

